Converts Rust i64 -> i32

+++ src/001/problem001.rs
@@ -1,4 +1,4 @@
-fn solve() -> i64 {
+fn solve() -> i32 {
     (0..1000).into_iter()
              .filter(|&x| x % 3 == 0 || x % 5 == 0)
              .sum()
